pyproject.toml
[tool.poetry]
name = "mkdocs-build-webhook"
version = "0.1.14"
description = ""
authors = ["RincewindWizzard <git@magierdinge.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.2"
loguru = "^0.7.2"
mkdocs = "^1.5.3"
python-dotenv = "^1.0.1"
gitpython = "^3.1.42"
pytest = "^8.1.1"
gunicorn = "^21.2.0"
toml = "^0.10.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
mkdocs-build-webhook = "mkdocs_build_webhook.__main__:main"
